Rename UserProfile redux folder to user, to be more general to other data besides profile
