+++ src/components/NavBar/NavBar.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import AppBar from 'material-ui/AppBar';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { actions as userActions } from '../../redux/UserProfile';
+import { actions as userActions } from '../../redux/user';
 import SearchBar from './SearchBar';
 import UserMenu from './UserMenu';
 import Login from './LoginButton';
