+++ src/redux/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import { reducer as userReducer } from './UserProfile';
 import { reducer as mapReducer } from './MapSearch';
+import { reducer as userReducer } from './user';
 
 const combinedReducer = combineReducers({
   userReducer,
